requirements.txt
flask
requests
arabic-reshaper
python-bidi
colorama
gunicorn
